+++ protocol-designer/src/components/organisms/ConfirmDeleteEntityInUseModal/index.tsx
@@ -17,12 +17,11 @@ import { getMainPagePortalEl } from '../Portal'
 interface ConfirmDeleteEntityInUseModalProps {
   onClose: () => void
   onConfirm: () => void
-  type: 'clear' | 'reconfigure'
 }
 export function ConfirmDeleteEntityInUseModal(
   props: ConfirmDeleteEntityInUseModalProps
 ): JSX.Element {
-  const { onClose, onConfirm, type } = props
+  const { onClose, onConfirm,} = props
   const { t } = useTranslation(['onboarding', 'shared'])
 
   return createPortal(
@@ -30,7 +29,7 @@ export function ConfirmDeleteEntityInUseModal(
       <Modal
         marginLeft="0"
         zIndexOverlay={11}
-        title={t(`are_you_sure_${type}_slot`)}
+        title={t(`are_you_sure_clear_slot`)}
         type="warning"
         onClose={onClose}
         footer={
