+++ protocol-designer/src/components/organisms/index.ts
@@ -27,6 +27,7 @@ export * from './Navigation'
 export * from './Ot2Modules'
 export * from './PipetteInfoItem'
 export * from './Portal'
+export * from './SelectLabwareModal'
 export * from './SelectPipetteModal'
 export * from './SelectWellsModal'
 export * from './SettingsIcon'
