+++ protocol-designer/src/pages/Designer/DeckSetup/SlotOverflowMenu.tsx
@@ -246,7 +246,6 @@ export function SlotOverflowMenu(
           onClose={() => {
             setShowDeleteEntityInUseModal(false)
           }}
-          type="clear"
         />
       ) : null}
       {showNickNameModal && nickNameId != null ? (
